requirements.txt
anyio==4.12.0
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.11.12
cffi==2.0.0
cssselect2==0.8.0
defusedxml==0.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
pillow==12.0.0
pycparser==2.23
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-telegram-bot==22.5
rich==14.2.0
six==1.17.0
tinycss2==1.5.1
webencodings==0.5.1
